requirements_gpu.txt
tensorflow-gpu==1.15
scipy
scikit-learn